# Replaced slow Scanner by LineReader (own implementation)