--- a/src/main/java/org/asteriskjava/manager/internal/ManagerConnectionImpl.java
+++ b/src/main/java/org/asteriskjava/manager/internal/ManagerConnectionImpl.java
@@ -670,7 +670,7 @@ public class ManagerConnectionImpl implements ManagerConnection, Dispatcher
             if (Asterisk14outputPresent)
             {
                 List<String> outputList = Arrays
-                        .asList(showVersionFilesResponse.getOutput().split(SocketConnectionFacadeImpl.NL_PATTERN.pattern()));
+                        .asList(showVersionFilesResponse.getOutput().split(Pattern.quote(SocketConnectionFacadeImpl.NL_PATTERN)));
                 showVersionFilesResult = outputList;
             }
             else