When we did a test with heavy load (and tons of events), it turned out that the `Scanner` used in `SocketConnectionFacadeImpl` is slow and burns CPU.

I first removed the `Scanner` and used the underlaying `BufferedReader`, but this broke the detection of line endings.
So I ended up with replacing the combination of `BufferedReader` and `Scanner` by an own implementation called `LineReader`. It behaves like a `BufferedReader` except for a custom pattern for line endings.

BTW: for the implementation, I also had to migrate the patterns from regex to string

I did some benchmarks of how long it takes to slice 1GB of text into lines (see `LineReaderTest`):
13140ms	BufferedReader+Scanner
1768ms	LineReader
1678ms	BufferedReader