@seanbright 2018-01-31
> I first removed the Scanner and used the underlaying BufferedReader, but this broke the detection of line endings.

How so? BufferedReader.readLine() is documented to handle the 3 common EOL variants.