@seanbright 2018-01-31
Also, out of curiosity, can you run a benchmark using the `Scanner` + `InputStreamReader`? For example:

`this.scanner = new Scanner(new InputStreamReader(inputStream, encoding));`

`Scanner` is going to do it's own buffering, so maybe having the `BufferedReader` is adding unnecessary overhead.