@bsutton 2018-01-31
Do you need to do the benchmark using a socket. I wonder if performance
would be different.

Also if we are here, have you looked nio classes. They are meant to offer
much better performance.

On 1 Feb. 2018 4:03 am, "Sean Bright" <notifications@github.com> wrote:

> Also, out of curiosity, can you run a benchmark using the Scanner +
> InputStreamReader? For example:
>
> this.scanner = new Scanner(new InputStreamReader(inputStream, encoding));
>
> Scanner is going to do it's own buffering, so maybe having the
> BufferedReader is adding unnecessary overhead.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/asterisk-java/asterisk-java/pull/231#issuecomment-361999890>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AA29OPm8l_W_Z5J4A7YRDhBSXfQzaQPNks5tQJz5gaJpZM4R0NQC>
> .
>