@elara-leitstellentechnik 2018-02-01
Hi again,

@seanbright:
The Scanner exists, because Asterisk uses different line breaks at different places of manager events:
* normal lines are separated by `\r\n` and usually contain a key/value pair,
* in the `CommandResponse` (for a `CommandAction`), there is a special line containing the (multi-line) answer of the command separated by `\n`.
To distinguish this correctly, lines have to be separated by `\r\n` only. So the fast `BufferedReader` can't be used here and this is why there is a `Scanner`.

Unfortunately, I can't post an example (as my browser runs in a restrictive sandbox). So I can only point you to  `ManagerReaderImpl`, line 181: `Special handling for "Response: Follows" (CommandResponse)`.

Futher, I did benchmarks over network. This went well for the `Scanner`, but the `BufferedReader` and `LineReader` were faster than the 1Gbps connection. The results looked something like this:
* with Scanner: ~32MB/s and 12% CPU (one core of the quadcore)
* without Scanner: ~52MB/s and 4% CPU (network limitation)

The Scanner without BufferedReader was a bit slower:
13958ms InputStreamReader(w/o BufferedReader)+Scanner


@bsutton:
In fact, our code use NIO socket (just some hacky replacement). But performance issues where the reason why we changed. The normal socket is missing some timeout features that were crucial for our application.
I think NIO sockets perform much better if you are dealing with lots of connections opening and closing and everything is running in parallel. This does not seam to be the typical use case for AMI connections.